- Further class naming work and renaming
