+++ WaywardHorizonsWPF/WaywardHorizons/App.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using System.Windows;
-using WaywardHorizons.Helpers;
+using WaywardHorizons.Services;
 using WaywardHorizons.Interfaces;
 
 namespace WaywardHorizons
