+++ WaywardHorizonsWPF/WaywardHorizons/ViewModels/MenuViewModel.xaml.cs
@@ -3,7 +3,7 @@ using System.Windows.Controls;
 using System.Windows.Input;
 using System.Windows.Navigation;
 using WaywardHorizons.Characters;
-using WaywardHorizons.Helpers;
+using WaywardHorizons.Interfaces;
 
 namespace WaywardHorizons
 {
